move base to own repo, build client and server on that base
In order for this workflow to work with forks, the base image must be
in its own git repo.  Otherwise the client and server images will
always be pointing to the base image from fdroid, not the fork,
e.g. eighthave.

client and server need to use base from its own repo, and it should
use a versioned path to base so that new runs of these are
reproducible.